requirements.txt
anthropic
openai
slack-sdk
python-dotenv
pyyaml
requests
